+++ core-web/libs/sdk/react/src/lib/next/components/Container/Container.tsx
@@ -1,6 +1,6 @@
 import { useContext, useMemo } from 'react';
 
-import { ContainerNoFound, EmptyContainer } from './ContainerFallbakcs';
+import { ContainerNoFound, EmptyContainer } from './ContainerFallbacks';
 
 import { DotCMSPageContext } from '../../contexts/DotCMSPageContext';
 import { DotCMSColumnContainer, DotCMSContentlet } from '../../types';
